+++ source/widgets/options/stream/stream_model.h
@@ -12,8 +12,8 @@ class StreamModel : public AbstractModel {
         static const QString FIELD_BITS_PER_SAMPLE;
         static const QString FIELD_SIGNED;
         
-        quint32 get_SampleRate();
-        void set_SampleRate(quint32 value);        
+        quint64 get_SampleRate();
+        void set_SampleRate(quint64 value);
         quint8 get_BitsPerSample();
         void set_BitsPerSample(quint8 value);
         bool get_Signed();
